# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LQGovernance — Board Management Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings accessible to current user
      parameters:
        - name: boardId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a meeting (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingBody"
      responses:
        "201":
          description: Created meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting by ID with agenda
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Meeting with agenda
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
    patch:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeetingBody"
      responses:
        "200":
          description: Updated meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "409":
          description: Illegal state transition or edit of a non-scheduled/cancelled meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /meetings/{id}/attendance:
    get:
      operationId: getMeetingAttendance
      tags: [meetings]
      summary: Get attendance for a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attendance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    patch:
      operationId: updateAttendance
      tags: [meetings]
      summary: Update attendance status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceBody"
      responses:
        "200":
          description: Updated attendance

  # Votes
...
  /meetings/{id}/agenda:
    post:
      operationId: addAgendaItem
      tags: [meetings]
      summary: Append an agenda item to a meeting (admin only; position is assigned server-side)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAgendaItemBody"
      responses:
        "201":
          description: Created agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"
        "409":
          description: The meeting is cancelled and immutable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meetings/{id}/agenda/{itemId}:
    patch:
      operationId: updateAgendaItem
      tags: [meetings]
      summary: Update an agenda item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgendaItemBody"
      responses:
        "200":
          description: Updated agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"
        "404":
          description: Agenda item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAgendaItem
      tags: [meetings]
      summary: Delete an agenda item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Tasks: evidence submission ──
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
...
    Person:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member, observer, management]
        title:
          type: ["string", "null"]
        avatarColor:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Meeting:
      type: object
      properties:
        id:
          type: string
        boardId:
          type: string
        title:
          type: string
        date:
          type: string
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, concluded, cancelled]
        boardName:
          type: ["string", "null"]
        boardAbbreviation:
          type: ["string", "null"]
        agendaItemCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - boardId
        - title
        - date
        - status
        - agendaItemCount
        - createdAt
    MeetingDetail:
      type: object
      properties:
        id:
          type: string
        boardId:
          type: string
        title:
          type: string
        date:
          type: string
        location:
          type: ["string", "null"]
        status:
          type: string
        boardName:
          type: ["string", "null"]
        boardAbbreviation:
          type: ["string", "null"]
        agendaItems:
          type: array
          items:
            $ref: "#/components/schemas/AgendaItem"
        attendance:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        createdAt:
          type: string
      required:
        - id
        - boardId
        - title
        - date
        - status
        - agendaItems
        - attendance
        - createdAt
    AgendaItem:
      type: object
      properties:
        id:
          type: string
        meetingId:
          type: string
        position:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [information, discussion, decision]
        description:
          type: ["string", "null"]
        voteId:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required:
        - id
        - meetingId
        - position
        - title
        - type
        - documents
    AttendanceRecord:
      type: object
      properties:
        personId:
          type: string
        meetingId:
          type: string
        status:
          type: string
          enum: [confirmed, pending, proxy, absent]
        proxyHolderId:
          type: ["string", "null"]
        person:
          $ref: "#/components/schemas/Person"
      required:
        - personId
        - meetingId
        - status
        - person
    CreateMeetingBody:
      type: object
      properties:
        boardId:
          type: string
        title:
          type: string
        date:
          type: string
        location:
          type: ["string", "null"]
        agendaItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateAgendaItemBody"
      required:
        - boardId
        - title
        - date
    CreateAgendaItemBody:
      type: object
      properties:
        position:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [information, discussion, decision]
        description:
          type: ["string", "null"]
      required:
        - position
        - title
        - type
    UpdateMeetingBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [scheduled, concluded, cancelled, null]
          description: >
            Legal transitions: scheduled → concluded|cancelled, concluded →
            scheduled (reopen). cancelled is terminal — the record is kept
            (cancel, not delete). Content edits require a scheduled meeting.
    UpdateAttendanceBody:
      type: object
      properties:
        updates:
          type: array
          items:
            type: object
            properties:
              personId:
                type: string
              status:
                type: string
                enum: [confirmed, pending, proxy, absent]
              proxyHolderId:
                type: ["string", "null"]
            required:
              - personId
              - status
      required:
        - updates

    # Vote
...
    Document:
      type: object
      properties:
        id:
          type: string
        boardId:
          type: ["string", "null"]
        title:
          type: string
        filename:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        aiClassification:
          type: ["object", "null"]
        uploadedBy:
          type: ["string", "null"]
        uploaderName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - filename
        - createdAt

    # Task
...
    UpdateAgendaItemBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [information, discussion, decision]
        description:
          type: string
    AddAgendaItemBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [information, discussion, decision]
        description:
          type: string
      required:
        - title
        - type

    # ── Vote certificates and documents ──
...